+++ CHANGES.rst
@@ -35,6 +35,7 @@
 - Fixed a bug in ``TargetPixelFile.animate()`` which caused a ``ModuleNotFoundError``
   to be raised when using older versions of matplotlib. [#1139]
 
+- Added a ``column`` parameter to ``LightCurve.truncate()``. [#1116]
 
 
 2.0.10 (2021-06-04)